Move SByteTests
